requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
feedparser==6.0.11
httpx==0.27.2
python-dotenv==1.0.1
openai==1.51.0
aiosqlite==0.20.0
jinja2==3.1.4
python-multipart==0.0.12
